INTENTO_2/requirements.txt
# Core RAG dependencies
openai>=1.12.0
chromadb>=0.4.22
pymupdf>=1.23.0          # PyMuPDF for PDF extraction (import fitz)
tiktoken>=0.6.0          # Token counting / chunking

# Web interface
streamlit>=1.32.0
streamlit-chat>=0.1.1

# Data & utilities
python-dotenv>=1.0.0
numpy>=1.26.0
pandas>=2.2.0
tqdm>=4.66.0
pydantic>=2.6.0

# Jupyter notebooks
notebook>=7.1.0
ipykernel>=6.29.0
ipywidgets>=8.1.0

# Visualization
matplotlib>=3.8.0
plotly>=5.20.0

# Evaluation
scipy>=1.12.0
